python/requirements.txt

openai
python-dotenv
tiktoken